Removed various minor dependencies restricting the splitting of Graql
