+++ graql/java/exception/BUILD
@@ -26,10 +26,9 @@ java_library(
 
 load("//dependencies/tools/checkstyle:checkstyle.bzl", "checkstyle_test")
 checkstyle_test(
- name = "exception-checkstyle",
- target = ":exception",
- config = "//config/checkstyle:checkstyle.xml",
- suppressions = "//config/checkstyle:checkstyle-suppressions.xml",
- licenses = ["//config/checkstyle:licenses"],
-
+    name = "exception-checkstyle",
+    target = ":exception",
+    config = "//config/checkstyle:checkstyle.xml",
+    suppressions = "//config/checkstyle:checkstyle-suppressions.xml",
+    licenses = ["//config/checkstyle:licenses"],
 )
