+++ server/src/graql/query/ComputeQuery.java
@@ -28,7 +28,6 @@ import grakn.core.graql.answer.ConceptSetMeasure;
 import grakn.core.graql.answer.Value;
 import grakn.core.graql.concept.ConceptId;
 import grakn.core.graql.exception.GraqlQueryException;
-import grakn.core.graql.util.StringUtil;
 
 import javax.annotation.CheckReturnValue;
 import java.util.ArrayList;
@@ -46,6 +45,7 @@ import java.util.Set;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
+import static graql.util.StringUtil.escapeLabelOrId;
 import static java.util.stream.Collectors.joining;
 
 
@@ -399,11 +399,11 @@ public class ComputeQuery<T extends Answer> implements Query {
 
         if (!types.isEmpty()) {
             if (types.size() == 1) {
-                inTypesString.append(StringUtil.typeLabelToString(types.iterator().next()));
+                inTypesString.append(escapeLabelOrId(types.iterator().next()));
             } else {
                 inTypesString.append(Char.SQUARE_OPEN);
                 inTypesString.append(inTypes.stream()
-                                             .map(StringUtil::typeLabelToString)
+                                             .map(name -> escapeLabelOrId(name))
                                              .collect(joining(Char.COMMA_SPACE.toString())));
                 inTypesString.append(Char.SQUARE_CLOSE);
             }
