+++ server/src/graql/query/pattern/property/WhenProperty.java
@@ -23,8 +23,6 @@ import grakn.core.graql.query.pattern.Conjunction;
 import grakn.core.graql.query.pattern.Pattern;
 import grakn.core.graql.query.pattern.statement.StatementType;
 
-import java.util.stream.Collectors;
-
 import static java.util.stream.Collectors.joining;
 
 /**
