+++ server/src/graql/util/StringUtil.java
@@ -17,53 +17,19 @@
  */
 package grakn.core.graql.util;
 
-import com.google.common.collect.ImmutableSet;
 import grakn.core.graql.concept.ConceptId;
 import grakn.core.graql.concept.Label;
-import graql.grammar.GraqlLexer;
-import org.apache.commons.lang.StringEscapeUtils;
 
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
-import java.util.HashSet;
 import java.util.Locale;
-import java.util.Set;
+
+import static graql.util.StringUtil.escapeLabelOrId;
 
 /**
  * Some helper methods in dealing with strings in the context of Graql.
  */
 public class StringUtil {
-
-    // TODO: This is critical knowledge that is lost in a Util class.
-    //       Move it to important class along with other Syntax rules
-    private static final ImmutableSet<String> ALLOWED_ID_KEYWORDS = ImmutableSet.of(
-            "min", "max", "median", "mean", "std", "sum", "count", "path", "cluster", "degrees", "members", "persist"
-    );
-    public static final Set<String> GRAQL_KEYWORDS = getKeywords();
-
-    /**
-     * @param string the string to unescape
-     * @return the unescaped string, replacing any backslash escapes with the real characters
-     */
-    public static String unescapeString(String string) {
-        return StringEscapeUtils.unescapeJavaScript(string);
-    }
-
-    /**
-     * @param string the string to escape
-     * @return the escaped string, replacing any escapable characters with backslashes
-     */
-    public static String escapeString(String string) {
-        return StringEscapeUtils.escapeJavaScript(string);
-    }
-
-    /**
-     * @param string a string to quote and escape
-     * @return a string, surrounded with double quotes and escaped
-     */
-    public static String quoteString(String string) {
-        return "\"" + StringUtil.escapeString(string) + "\"";
-    }
 
     /**
      * @param value a value in the graph
@@ -71,7 +37,7 @@ public class StringUtil {
      */
     public static String valueToString(Object value) {
         if (value instanceof String) {
-            return quoteString((String) value);
+            return graql.util.StringUtil.quoteString((String) value);
         } else if (value instanceof Double) {
             DecimalFormat df = new DecimalFormat("#", DecimalFormatSymbols.getInstance(Locale.ENGLISH));
             df.setMinimumFractionDigits(1);
@@ -103,34 +69,5 @@ public class StringUtil {
      */
     public static String typeLabelToString(Label label) {
         return escapeLabelOrId(label.getValue());
-    }
-
-    public static String typeLabelToString(String name) {
-        return escapeLabelOrId(name);
-    }
-
-    private static String escapeLabelOrId(String value) {
-        if (value.matches("^@?[a-zA-Z_][a-zA-Z0-9_-]*$") &&
-                (!GRAQL_KEYWORDS.contains(value)) || ALLOWED_ID_KEYWORDS.contains(value)) {
-            return value;
-        } else {
-            return quoteString(value);
-        }
-    }
-
-    /**
-     * @return all Graql keywords
-     */
-    private static Set<String> getKeywords() {
-        HashSet<String> keywords = new HashSet<>();
-
-        for (int i = 1; i <= GraqlLexer.VOCABULARY.getMaxTokenType(); i ++) {
-            if (GraqlLexer.VOCABULARY.getLiteralName(i) != null) {
-                String name = GraqlLexer.VOCABULARY.getLiteralName(i);
-                keywords.add(name.replaceAll("'", ""));
-            }
-        }
-
-        return ImmutableSet.copyOf(keywords);
     }
 }
